requirements.txt
flask==3.0.3
diffusers==0.31.0
transformers==4.47.1
accelerate==1.9.0
sentencepiece==0.2.0
pillow==11.0.0
python-dotenv==1.0.1
huggingface-hub==0.34.3
safetensors==0.5.3
requests==2.32.3